+++ pkg/monitortests/kubeapiserver/disruptioninclusterapiserver/manifests/role-monitor.yaml
