--- pkg/monitortests/kubeapiserver/incluster_disruption_serializer/manifests/dep-service-network.yaml
@@ -1,56 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: service-network-monitor
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: service-network-monitor
-      network.openshift.io/disruption-actor: poller
-  template:
-    metadata:
-      labels:
-        app: service-network-monitor
-        network.openshift.io/disruption-actor: poller
-    spec:
-      affinity:
-        podAntiAffinity:
-          requiredDuringSchedulingIgnoredDuringExecution:
-          - labelSelector:
-              matchExpressions:
-              - key: app
-                operator: In
-                values:
-                - service-network-monitor
-            topologyKey: "kubernetes.io/hostname"
-      containers:
-      - name: service-network
-        command:
-        - openshift-tests
-        - run-disruption
-        - --output-file=/var/log/disruption-data/monitor-events/service-network-monitor.txt
-        - --lb-type=$(LOAD_BALANCER)
-        - --stop-configmap=stop-configmap
-        env:
-        - name: LOAD_BALANCER
-          value: service-network
-        image: "image-registry.openshift-image-registry.svc:5000/openshift/tests:latest"
-        volumeMounts:
-        - mountPath: /var/log/disruption-data
-          name: artifacts
-        securityContext:
-          privileged: true
-          runAsUser: 0
-      serviceAccountName: disruption-monitor-sa
-      securityContext:
-        privileged: true
-        runAsUser: 0
-      volumes:
-      - hostPath:
-          path: /var/log/disruption-data
-          type: DirectoryOrCreate
-        name: artifacts
-      tolerations:
-      - key: node-role.kubernetes.io/master
-        effect: NoSchedule
